requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sentence-transformers
asyncpg
psycopg2-binary
pgvector
sqlalchemy[asyncio]
alembic
tree-sitter==0.21.3
tree-sitter-python==0.21.0
rank-bm25
dramatiq[redis]
redis
httpx
python-dotenv
ragas
groq
websockets
shortuuid